refactor scattergl - create hover and select files
