+++ src/traces/scattergl/index.js
@@ -8,258 +8,7 @@
 
 'use strict';
 
-var Registry = require('../../registry');
-var Lib = require('../../lib');
-
-var subTypes = require('../scatter/subtypes');
-var getTraceColor = require('../scatter/get_trace_color');
-var fillText = Lib.fillText;
-
-var styleTextSelection = require('./style').styleTextSelection;
-
-function hoverPoints(pointData, xval, yval, hovermode) {
-    var cd = pointData.cd;
-    var stash = cd[0].t;
-    var trace = cd[0].trace;
-    var xa = pointData.xa;
-    var ya = pointData.ya;
-    var x = stash.x;
-    var y = stash.y;
-    var xpx = xa.c2p(xval);
-    var ypx = ya.c2p(yval);
-    var maxDistance = pointData.distance;
-    var ids;
-
-    // FIXME: make sure this is a proper way to calc search radius
-    if(stash.tree) {
-        var xl = xa.p2c(xpx - maxDistance);
-        var xr = xa.p2c(xpx + maxDistance);
-        var yl = ya.p2c(ypx - maxDistance);
-        var yr = ya.p2c(ypx + maxDistance);
-
-        if(hovermode === 'x') {
-            ids = stash.tree.range(
-                Math.min(xl, xr), Math.min(ya._rl[0], ya._rl[1]),
-                Math.max(xl, xr), Math.max(ya._rl[0], ya._rl[1])
-            );
-        } else {
-            ids = stash.tree.range(
-                Math.min(xl, xr), Math.min(yl, yr),
-                Math.max(xl, xr), Math.max(yl, yr)
-            );
-        }
-    } else if(stash.ids) {
-        ids = stash.ids;
-    } else return [pointData];
-
-    // pick the id closest to the point
-    // note that point possibly may not be found
-    var id, ptx, pty, i, dx, dy, dist, dxy;
-
-    var minDist = maxDistance;
-    if(hovermode === 'x') {
-        for(i = 0; i < ids.length; i++) {
-            ptx = x[ids[i]];
-            dx = Math.abs(xa.c2p(ptx) - xpx);
-            if(dx < minDist) {
-                minDist = dx;
-                dy = ya.c2p(y[ids[i]]) - ypx;
-                dxy = Math.sqrt(dx * dx + dy * dy);
-                id = ids[i];
-            }
-        }
-    } else {
-        for(i = ids.length - 1; i > -1; i--) {
-            ptx = x[ids[i]];
-            pty = y[ids[i]];
-            dx = xa.c2p(ptx) - xpx;
-            dy = ya.c2p(pty) - ypx;
-
-            dist = Math.sqrt(dx * dx + dy * dy);
-            if(dist < minDist) {
-                minDist = dxy = dist;
-                id = ids[i];
-            }
-        }
-    }
-
-    pointData.index = id;
-    pointData.distance = minDist;
-    pointData.dxy = dxy;
-
-    if(id === undefined) return [pointData];
-
-    calcHover(pointData, x, y, trace);
-
-    return [pointData];
-}
-
-function calcHover(pointData, x, y, trace) {
-    var xa = pointData.xa;
-    var ya = pointData.ya;
-    var minDist = pointData.distance;
-    var dxy = pointData.dxy;
-    var id = pointData.index;
-
-    // the closest data point
-    var di = {
-        pointNumber: id,
-        x: x[id],
-        y: y[id]
-    };
-
-    // that is single-item arrays_to_calcdata excerpt, since we are doing it for a single point and we don't have to do it beforehead for 1e6 points
-    di.tx = Array.isArray(trace.text) ? trace.text[id] : trace.text;
-    di.htx = Array.isArray(trace.hovertext) ? trace.hovertext[id] : trace.hovertext;
-    di.data = Array.isArray(trace.customdata) ? trace.customdata[id] : trace.customdata;
-    di.tp = Array.isArray(trace.textposition) ? trace.textposition[id] : trace.textposition;
-
-    var font = trace.textfont;
-    if(font) {
-        di.ts = Array.isArray(font.size) ? font.size[id] : font.size;
-        di.tc = Array.isArray(font.color) ? font.color[id] : font.color;
-        di.tf = Array.isArray(font.family) ? font.family[id] : font.family;
-    }
-
-    var marker = trace.marker;
-    if(marker) {
-        di.ms = Lib.isArrayOrTypedArray(marker.size) ? marker.size[id] : marker.size;
-        di.mo = Lib.isArrayOrTypedArray(marker.opacity) ? marker.opacity[id] : marker.opacity;
-        di.mx = Array.isArray(marker.symbol) ? marker.symbol[id] : marker.symbol;
-        di.mc = Lib.isArrayOrTypedArray(marker.color) ? marker.color[id] : marker.color;
-    }
-
-    var line = marker && marker.line;
-    if(line) {
-        di.mlc = Array.isArray(line.color) ? line.color[id] : line.color;
-        di.mlw = Lib.isArrayOrTypedArray(line.width) ? line.width[id] : line.width;
-    }
-
-    var grad = marker && marker.gradient;
-    if(grad && grad.type !== 'none') {
-        di.mgt = Array.isArray(grad.type) ? grad.type[id] : grad.type;
-        di.mgc = Array.isArray(grad.color) ? grad.color[id] : grad.color;
-    }
-
-    var xp = xa.c2p(di.x, true);
-    var yp = ya.c2p(di.y, true);
-    var rad = di.mrc || 1;
-
-    var hoverlabel = trace.hoverlabel;
-
-    if(hoverlabel) {
-        di.hbg = Array.isArray(hoverlabel.bgcolor) ? hoverlabel.bgcolor[id] : hoverlabel.bgcolor;
-        di.hbc = Array.isArray(hoverlabel.bordercolor) ? hoverlabel.bordercolor[id] : hoverlabel.bordercolor;
-        di.hts = Array.isArray(hoverlabel.font.size) ? hoverlabel.font.size[id] : hoverlabel.font.size;
-        di.htc = Array.isArray(hoverlabel.font.color) ? hoverlabel.font.color[id] : hoverlabel.font.color;
-        di.htf = Array.isArray(hoverlabel.font.family) ? hoverlabel.font.family[id] : hoverlabel.font.family;
-        di.hnl = Array.isArray(hoverlabel.namelength) ? hoverlabel.namelength[id] : hoverlabel.namelength;
-    }
-    var hoverinfo = trace.hoverinfo;
-    if(hoverinfo) {
-        di.hi = Array.isArray(hoverinfo) ? hoverinfo[id] : hoverinfo;
-    }
-
-    var hovertemplate = trace.hovertemplate;
-    if(hovertemplate) {
-        di.ht = Array.isArray(hovertemplate) ? hovertemplate[id] : hovertemplate;
-    }
-
-    var fakeCd = {};
-    fakeCd[pointData.index] = di;
-
-    Lib.extendFlat(pointData, {
-        color: getTraceColor(trace, di),
-
-        x0: xp - rad,
-        x1: xp + rad,
-        xLabelVal: di.x,
-
-        y0: yp - rad,
-        y1: yp + rad,
-        yLabelVal: di.y,
-
-        cd: fakeCd,
-        distance: minDist,
-        spikeDistance: dxy,
-
-        hovertemplate: di.ht
-    });
-
-    if(di.htx) pointData.text = di.htx;
-    else if(di.tx) pointData.text = di.tx;
-    else if(trace.text) pointData.text = trace.text;
-
-    fillText(di, trace, pointData);
-    Registry.getComponentMethod('errorbars', 'hoverInfo')(di, trace, pointData);
-
-    return pointData;
-}
-
-function selectPoints(searchInfo, selectionTester) {
-    var cd = searchInfo.cd;
-    var selection = [];
-    var trace = cd[0].trace;
-    var stash = cd[0].t;
-    var len = trace._length;
-    var x = stash.x;
-    var y = stash.y;
-    var scene = stash._scene;
-    var index = stash.index;
-
-    if(!scene) return selection;
-
-    var hasText = subTypes.hasText(trace);
-    var hasMarkers = subTypes.hasMarkers(trace);
-    var hasOnlyLines = !hasMarkers && !hasText;
-
-    if(trace.visible !== true || hasOnlyLines) return selection;
-
-    var els = [];
-    var unels = [];
-
-    // degenerate polygon does not enable selection
-    // filter out points by visible scatter ones
-    if(selectionTester !== false && !selectionTester.degenerate) {
-        for(var i = 0; i < len; i++) {
-            if(selectionTester.contains([stash.xpx[i], stash.ypx[i]], false, i, searchInfo)) {
-                els.push(i);
-                selection.push({
-                    pointNumber: i,
-                    x: x[i],
-                    y: y[i]
-                });
-            } else {
-                unels.push(i);
-            }
-        }
-    }
-
-    if(hasMarkers) {
-        var scatter2d = scene.scatter2d;
-
-        if(!els.length && !unels.length) {
-            // reset to base styles when clearing
-            var baseOpts = new Array(scene.count);
-            baseOpts[index] = scene.markerOptions[index];
-            scatter2d.update.apply(scatter2d, baseOpts);
-        } else if(!scene.selectBatch[index].length && !scene.unselectBatch[index].length) {
-            // set unselected styles on 'context' canvas (if not done already)
-            var unselOpts = new Array(scene.count);
-            unselOpts[index] = scene.markerUnselectedOptions[index];
-            scatter2d.update.apply(scatter2d, unselOpts);
-        }
-    }
-
-    scene.selectBatch[index] = els;
-    scene.unselectBatch[index] = unels;
-
-    if(hasText) {
-        styleTextSelection(cd);
-    }
-
-    return selection;
-}
+var hover = require('./hover');
 
 module.exports = {
     moduleType: 'trace',
@@ -273,11 +22,11 @@ module.exports = {
     colorbar: require('../scatter/marker_colorbar'),
     calc: require('./calc'),
     plot: require('./plot'),
-    hoverPoints: hoverPoints,
-    selectPoints: selectPoints,
+    hoverPoints: hover.hoverPoints,
+    selectPoints: require('./select'),
 
     sceneUpdate: require('./scene_update'),
-    calcHover: calcHover,
+    calcHover: hover.calcHover,
 
     meta: {
         hrName: 'scatter_gl',
